refactor(cli): move build handler to `cli-functions`
